feat(cf-deploy-config): sonar coverage
